refactor(eval): inline oracle sweep into EvalS1Task; own AI subtasks
The eval tasks split their logic into thin task wrappers around free
functions instead of living inside the task, as the libs task idiom
intends. EvalS1Task delegated its entire body to runOracleEval, which was
imported by nothing else and had no test of its own.

- Inline runOracleEval (the oracle comparison over real committed S-1
  sections) directly into EvalS1Task.execute, moving its helpers
  (runSection, summarize) and report types (OracleReport, OracleRunResult,
  OracleModelSummary, GOLDEN_REFERENCE) into the task file. The task now
  drives progress through its own execute context rather than an onProgress
  shim, and prefetches models through the real context. Delete the now-dead
  runOracleEval.ts; the CLI imports OracleReport from the task.

- Own the AI subtasks the extraction path instantiates. runStructured
  (StructuredGenerationTask) and ensureModelDownloaded (ModelDownloadTask)
  now wrap the instance in context.own(...), so when a task threads its real
  execute context down (form processing, the eval prefetch) the subtask is
  registered in that task's graph and inherits its registry + abort signal.
  Against the deliberate one-shot stub context (eval sweeps, unit tests),
  own is an identity no-op, so behavior there is unchanged.

The sibling run functions (runExtractionEval, runUnitTermsEval) are left as
standalone: they are unit-tested directly and export helpers consumed
elsewhere, so inlining them would break tests and duplicate shared code.
